Update main.c
This program contains function to add a node at the end of the list, to dequeue the last element from the list, insert a node at a particular position, delete an nth node 
and reverse the linked list. And of course, printing the output.

The main focus was implementing the linked list. The size (to print) increments even if the input is INVALID. But it can be fixed by adding some additional loops. Printing statements can be improved.